requirements.txt
boto3
pymongo
requests
pyyaml
python-dotenv
